pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = []

[project]
name = "gridfinityworkbench"
version = "0.9.5"

[project.optional-dependencies]
dev = ["ruff==0.9.5", "freecad-stubs==1.0.18"]

[tool.ruff]
line-length = 100
extend-exclude = ["freecad/gridfinity_workbench/preferences/auto.py"]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["EM101", "EM102", "RET504", "TRY003"]

[tool.ruff.lint.per-file-ignores]
# Ignore folowing rules for tests:
# INP001 - implicit-namespace-package: no __init__ is needed for test files
# D100   - undocumented-public-module
# D101   - undocumented-public-class
# D102   - undocumented-public-method
# PT027  - pytest-unittest-raises-assertion: we use unittest framework, not pytest
# PT009  - pytest-unittest-assertion: we use unittest framework, not pytest
"**/tests/*" = ["INP001", "D100", "D101", "D102", "PT027", "PT009"]
